modify

+++ ccache.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 version=3.5
-rm -rf ccache-${version}.tar.gz
-wget https://www.samba.org/ftp/ccache/ccache-${version}.tar.gz
-rm -rf ccache-${version}
-tar -xf ccache-${version}.tar.gz
+if [ ! -f ccache-${version}.tar.gz ]
+then
+    wget https://www.samba.org/ftp/ccache/ccache-${version}.tar.gz
+    tar -xf ccache-${version}.tar.gz
+fi
 cd ccache-${version}
-./configure
-sudo make install
-cd ..
-rm -rf ccache-${version}.tar.gz ccache-${version}
+./configure --prefix=${HOME}/usr
+make install
+if [ 0 -eq $? ]
+then
+    cd ..
+    rm -rf ccache-${version}.tar.gz ccache-${version}
+fi
